+++ modules/tinymce/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed the deprecated `$`, `DomQuery` and `Sizzle` APIs #TINY-4520
 - Removed the deprecated `Color`, `JSON`, `JSONP` and `JSONRequest` #TINY-8162
 - Removed the legacy browser detection properties from `Env` #TINY-8162
+- Removed the deprecated `setIconStroke` Split Toolbar Button API #TINY-8162
 - Removed the deprecated `editors` property from `EditorManager` #TINY-8162
 - Removed the deprecated `execCallback` and `setMode` APIs from `Editor` #TINY-8162
 - Removed the deprecated `addComponents` and `dependencies` APIs from `AddOnManager` #TINY-8162